refactor(backend): isolate direct evidence prompts
- Move live-evidence planner contract and hint-list formatting into direct_prompt_evidence.py

- Keep direct_prompts.py focused on prompt assembly while preserving source-backed guidance text

- Update runtime cleanup audit with the #537 boundary

Refs #537
